Refactored fuctions
